pyproject.toml
[tool.poetry]
name = "conceptnet-rocks"
version = "0.1.0"
description = "Python library to work with ConceptNet offline"
authors = ["Roman Inflianskas <infroma@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
python-arango = "^6.0.0"
typer = "^0.3.2"
tqdm = "^4.48.2"
pySmartDL = "^1.3.3"
circus = "^0.16.1"
psutil = "^5.7.2"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
conceptnet-rocks = "conceptnet_rocks.cli:app"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
